fix(mge/optimizer): allow lr to be 0
GitOrigin-RevId: dabd1fcc3390787988553abded8c572bfccf1957

+++ imperative/python/megengine/optimizer/adagrad.py
@@ -62,7 +62,7 @@ class Adagrad(Optimizer):
         eps = param_group["eps"]
 
         def make_scalar(val):
-            return tensor(val)
+            return tensor(val, dtype="float32")
 
         # since `conver_inputs` is disabled for param updates,
         # scalar should be explicitly tansforred to tensor
